Remove backspace, fix operator state bug, add double-zero
Backspace caused issues that were not worth the trouble to
resolve. It was an experiment anyway.

Added double-zero button in place of backspace, removed
type conversion of all numerals to Numbers since it didn't
seem to affect the functionality and was preventing double-
zero from working since it would resolve to 0.

Operator states would not be deactivated when a second
operator was clicked in a row. Removed switch statement and
now we remove all active states and add back the newly active
one.